+++ core/amber/src/main/scala/edu/uci/ics/texera/workflow/common/workflow/LogicalLink.scala
@@ -1,25 +1,3 @@
 package edu.uci.ics.texera.workflow.common.workflow
 
-import edu.uci.ics.amber.engine.common.virtualidentity.OperatorIdentity
-
-case object LogicalPort {
-  def apply(operatorIdentity: OperatorIdentity, portOrdinal: Integer): LogicalPort = {
-    LogicalPort(operatorIdentity.id, portOrdinal)
-  }
-  def apply(
-      operatorIdentity: OperatorIdentity,
-      portOrdinal: Integer,
-      portName: String
-  ): LogicalPort = {
-    LogicalPort(operatorIdentity.id, portOrdinal, portName)
-  }
-}
-case class LogicalPort(
-    operatorID: String,
-    portOrdinal: Integer = 0,
-    portName: String = ""
-) {
-  def operatorId: OperatorIdentity = OperatorIdentity(operatorID)
-}
-
 case class LogicalLink(origin: LogicalPort, destination: LogicalPort)
